+++ QueueIntegrations/Prognetics.Saga.Queue.RabbitMQ/ChannelSetup/IRabbitMqChannelFactory.cs
@@ -1,0 +1,8 @@
+﻿using RabbitMQ.Client;
+
+namespace Prognetics.Saga.Queue.RabbitMQ.ChannelSetup;
+
+public interface IRabbitMqChannelFactory
+{
+    IModel Create();
+}
