+++ QueueIntegrations/Prognetics.Saga.Queue.RabbitMQ/ChannelSetup/IRabbitMqSagaQueuesProvider.cs
@@ -1,0 +1,6 @@
+﻿namespace Prognetics.Saga.Queue.RabbitMQ.ChannelSetup;
+
+public interface IRabbitMqSagaQueuesProvider
+{
+    public IReadOnlyList<RabbitMqQueue> Queues { get; }
+}
