+++ QueueIntegrations/Prognetics.Saga.Queue.RabbitMQ/Consuming/IRabbitMqSagaConsumerFactory.cs
@@ -1,0 +1,11 @@
+﻿using Prognetics.Saga.Orchestrator;
+using RabbitMQ.Client;
+
+namespace Prognetics.Saga.Queue.RabbitMQ.Consuming;
+
+internal interface IRabbitMqSagaConsumerFactory
+{
+    IBasicConsumer Create(
+        IModel channel,
+        ISagaQueue sagaQueue);
+}
