+++ QueueIntegrations/Prognetics.Saga.Queue.RabbitMQ/Consuming/IRabbitMqSagaConsumersFactory.cs
@@ -1,0 +1,11 @@
+﻿using Prognetics.Saga.Orchestrator;
+using RabbitMQ.Client;
+
+namespace Prognetics.Saga.Queue.RabbitMQ.Consuming;
+
+public interface IRabbitMqSagaConsumersFactory
+{
+    IReadOnlyList<RabbitMqConsumer> Create(
+        IModel channel,
+        ISagaQueue sagaQueue);
+}
