+++ QueueIntegrations/Prognetics.Saga.Queue.RabbitMQ/Hosting/IRabbitMqSagaHostingService.cs
@@ -1,0 +1,6 @@
+﻿namespace Prognetics.Saga.Queue.RabbitMQ.Hosting;
+
+public interface IRabbitMqSagaHostingService
+{
+    Task Listen(CancellationToken cancellationToken);
+}
