+++ QueueIntegrations/Prognetics.Saga.Queue.RabbitMQ/Serialization/IRabbitMqSagaSerializer.cs
@@ -1,0 +1,10 @@
+﻿using Prognetics.Saga.Orchestrator;
+using System;
+
+namespace Prognetics.Saga.Queue.RabbitMQ.Serialization;
+
+public interface IRabbitMqSagaSerializer
+{
+    byte[] Serialize(OutputMessage inputMessage);
+    InputMessage? Deserialize(ReadOnlyMemory<byte> messageBytes);
+}
