+++ QueueIntegrations/Prognetics.Saga.Queue.RabbitMQ/Subscribing/IRabbitMqSagaSubscriberFactory.cs
@@ -1,0 +1,9 @@
+﻿using Prognetics.Saga.Orchestrator;
+using RabbitMQ.Client;
+
+namespace Prognetics.Saga.Queue.RabbitMQ.Subscribing;
+
+public interface IRabbitMqSagaSubscriberFactory
+{
+    ISagaSubscriber Create(IModel model);
+}
